refactor: move migration to version 48
