+++ imports/plugins/core/versions/server/util/convert48.js
